[SIDE-2331] Refactoring
